+++ src/interfaces/IMotorcycle.ts
@@ -3,7 +3,7 @@ import { VehicleZodSchema } from './IVehicle';
 
 export const MotorcycleZodSchema = z.object({
   category: z.enum(['Street', 'Custom', 'Trail']),
-  engineCapacity: z.number().positive().max(2500),
+  engineCapacity: z.number().int().positive().max(2500),
 }).merge(VehicleZodSchema);
 
 export type IMotorcycle = z.infer<typeof MotorcycleZodSchema>;
